ui: extract agent tool policy helpers to keep files maintainable
