+++ ui/src/ui/views/agents.ts
@@ -11,12 +11,8 @@ import type {
   SkillStatusEntry,
   SkillStatusReport,
 } from "../types.ts";
-import { SUBAGENT_DEFAULT_TOOL_DENY } from "../../../../src/agents/subagent-tool-deny.js";
-import {
-  expandToolGroups,
-  normalizeToolName,
-  resolveToolProfilePolicy,
-} from "../../../../src/agents/tool-policy.js";
+import type { ConfigSnapshot, ToolPolicy } from "./agents-tool-policy.ts";
+import { normalizeToolName, resolveToolProfilePolicy } from "../../../../src/agents/tool-policy.js";
 import { formatRelativeTimestamp } from "../format.ts";
 import {
   formatCronPayload,
@@ -24,6 +20,14 @@ import {
   formatCronState,
   formatNextRun,
 } from "../presenter.ts";
+import {
+  PROFILE_OPTIONS,
+  SUBAGENT_LOCKED_TOOL_IDS,
+  TOOL_SECTIONS,
+  isAllowedByPolicy,
+  matchesList,
+  resolveAgentConfig,
+} from "./agents-tool-policy.ts";
 
 export type AgentsPanel = "overview" | "files" | "tools" | "skills" | "channels" | "cron";
 
@@ -85,141 +89,6 @@ export type AgentsProps = {
   onAgentSkillToggle: (agentId: string, skillName: string, enabled: boolean) => void;
   onAgentSkillsClear: (agentId: string) => void;
   onAgentSkillsDisableAll: (agentId: string) => void;
-};
-
-const TOOL_SECTIONS = [
-  {
-    id: "fs",
-    label: "Files",
-    tools: [
-      { id: "read", label: "read", description: "Read file contents" },
-      { id: "write", label: "write", description: "Create or overwrite files" },
-      { id: "edit", label: "edit", description: "Make precise edits" },
-      { id: "apply_patch", label: "apply_patch", description: "Patch files (OpenAI)" },
-    ],
-  },
-  {
-    id: "runtime",
-    label: "Runtime",
-    tools: [
-      { id: "exec", label: "exec", description: "Run shell commands" },
-      { id: "process", label: "process", description: "Manage background processes" },
-    ],
-  },
-  {
-    id: "web",
-    label: "Web",
-    tools: [
-      { id: "web_search", label: "web_search", description: "Search the web" },
-      { id: "web_fetch", label: "web_fetch", description: "Fetch web content" },
-    ],
-  },
-  {
-    id: "memory",
-    label: "Memory",
-    tools: [
-      { id: "memory_search", label: "memory_search", description: "Semantic search" },
-      { id: "memory_get", label: "memory_get", description: "Read memory files" },
-    ],
-  },
-  {
-    id: "sessions",
-    label: "Sessions",
-    tools: [
-      { id: "sessions_list", label: "sessions_list", description: "List sessions" },
-      { id: "sessions_history", label: "sessions_history", description: "Session history" },
-      { id: "sessions_send", label: "sessions_send", description: "Send to session" },
-      { id: "sessions_spawn", label: "sessions_spawn", description: "Spawn sub-agent" },
-      { id: "session_status", label: "session_status", description: "Session status" },
-    ],
-  },
-  {
-    id: "ui",
-    label: "UI",
-    tools: [
-      { id: "browser", label: "browser", description: "Control web browser" },
-      { id: "canvas", label: "canvas", description: "Control canvases" },
-    ],
-  },
-  {
-    id: "messaging",
-    label: "Messaging",
-    tools: [{ id: "message", label: "message", description: "Send messages" }],
-  },
-  {
-    id: "automation",
-    label: "Automation",
-    tools: [
-      { id: "cron", label: "cron", description: "Schedule tasks" },
-      { id: "gateway", label: "gateway", description: "Gateway control" },
-    ],
-  },
-  {
-    id: "nodes",
-    label: "Nodes",
-    tools: [{ id: "nodes", label: "nodes", description: "Nodes + devices" }],
-  },
-  {
-    id: "agents",
-    label: "Agents",
-    tools: [{ id: "agents_list", label: "agents_list", description: "List agents" }],
-  },
-  {
-    id: "media",
-    label: "Media",
-    tools: [{ id: "image", label: "image", description: "Image understanding" }],
-  },
-];
-
-const PROFILE_OPTIONS = [
-  { id: "minimal", label: "Minimal" },
-  { id: "coding", label: "Coding" },
-  { id: "messaging", label: "Messaging" },
-  { id: "full", label: "Full" },
-] as const;
-
-const SUBAGENT_LOCKED_TOOL_IDS = new Set(
-  SUBAGENT_DEFAULT_TOOL_DENY.map((entry) => normalizeToolName(entry)),
-);
-
-type ToolPolicy = {
-  allow?: string[];
-  deny?: string[];
-};
-
-type AgentConfigEntry = {
-  id: string;
-  name?: string;
-  workspace?: string;
-  agentDir?: string;
-  model?: unknown;
-  skills?: string[];
-  tools?: {
-    profile?: string;
-    allow?: string[];
-    alsoAllow?: string[];
-    deny?: string[];
-  };
-  subagents?: {
-    tools?: {
-      allow?: string[];
-      alsoAllow?: string[];
-      deny?: string[];
-    };
-  };
-};
-
-type ConfigSnapshot = {
-  agents?: {
-    defaults?: { workspace?: string; model?: unknown; models?: Record<string, { alias?: string }> };
-    list?: AgentConfigEntry[];
-  };
-  tools?: {
-    profile?: string;
-    allow?: string[];
-    alsoAllow?: string[];
-    deny?: string[];
-  };
 };
 
 function normalizeAgentLabel(agent: { id: string; name?: string; identity?: { name?: string } }) {
@@ -292,17 +161,6 @@ function formatBytes(bytes?: number) {
     unitIndex += 1;
   }
   return `${size.toFixed(size < 10 ? 1 : 0)} ${units[unitIndex]}`;
-}
-
-function resolveAgentConfig(config: Record<string, unknown> | null, agentId: string) {
-  const cfg = config as ConfigSnapshot | null;
-  const list = cfg?.agents?.list ?? [];
-  const entry = list.find((agent) => agent?.id === agentId);
-  return {
-    entry,
-    defaults: cfg?.agents?.defaults,
-    globalTools: cfg?.tools,
-  };
 }
 
 type AgentContext = {
@@ -465,89 +323,6 @@ function buildModelOptions(configForm: Record<string, unknown> | null, current?:
     `;
   }
   return options.map((option) => html`<option value=${option.value}>${option.label}</option>`);
-}
-
-type CompiledPattern =
-  | { kind: "all" }
-  | { kind: "exact"; value: string }
-  | { kind: "regex"; value: RegExp };
-
-function compilePattern(pattern: string): CompiledPattern {
-  const normalized = normalizeToolName(pattern);
-  if (!normalized) {
-    return { kind: "exact", value: "" };
-  }
-  if (normalized === "*") {
-    return { kind: "all" };
-  }
-  if (!normalized.includes("*")) {
-    return { kind: "exact", value: normalized };
-  }
-  const escaped = normalized.replace(/[.*+?^${}()|[\\]\\]/g, "\\$&");
-  return { kind: "regex", value: new RegExp(`^${escaped.replaceAll("\\*", ".*")}$`) };
-}
-
-function compilePatterns(patterns?: string[]): CompiledPattern[] {
-  if (!Array.isArray(patterns)) {
-    return [];
-  }
-  return expandToolGroups(patterns)
-    .map(compilePattern)
-    .filter((pattern) => {
-      return pattern.kind !== "exact" || pattern.value.length > 0;
-    });
-}
-
-function matchesAny(name: string, patterns: CompiledPattern[]) {
-  for (const pattern of patterns) {
-    if (pattern.kind === "all") {
-      return true;
-    }
-    if (pattern.kind === "exact" && name === pattern.value) {
-      return true;
-    }
-    if (pattern.kind === "regex" && pattern.value.test(name)) {
-      return true;
-    }
-  }
-  return false;
-}
-
-function isAllowedByPolicy(name: string, policy?: ToolPolicy) {
-  if (!policy) {
-    return true;
-  }
-  const normalized = normalizeToolName(name);
-  const deny = compilePatterns(policy.deny);
-  if (matchesAny(normalized, deny)) {
-    return false;
-  }
-  const allow = compilePatterns(policy.allow);
-  if (allow.length === 0) {
-    return true;
-  }
-  if (matchesAny(normalized, allow)) {
-    return true;
-  }
-  if (normalized === "apply_patch" && matchesAny("exec", allow)) {
-    return true;
-  }
-  return false;
-}
-
-function matchesList(name: string, list?: string[]) {
-  if (!Array.isArray(list) || list.length === 0) {
-    return false;
-  }
-  const normalized = normalizeToolName(name);
-  const patterns = compilePatterns(list);
-  if (matchesAny(normalized, patterns)) {
-    return true;
-  }
-  if (normalized === "apply_patch" && matchesAny("exec", patterns)) {
-    return true;
-  }
-  return false;
 }
 
 export function renderAgents(props: AgentsProps) {
